contrib/PaddleRendering/requirements.txt
colorlog
filelock
numpy
imageio==2.21.1
opencv-python
pyyaml
pillow
rarfile
visualdl
easydict
